move isoformat to compat, allow date, misc fixes
